Add stdint.h
